refactor: rename utils folder to scraper and update imports
