+++ scraper/scraper-functions.ts
